+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Improved chart rendering.
 * Improvements to container charts.
 * Created Supporters page.
+* Ability to remove images, volumes, or networks.
+* Updated templates.json.
 
 ## v0.20 (Jan 20th 2024) - The rewrite. Jumping all the way to v0.20.
 * Changed to ES6 imports.